+++ packages/patternfly-4/react-core/src/components/DataList/DataListAction.d.ts
@@ -1,9 +1,10 @@
-import { FunctionComponent, HTMLProps } from 'react';
+import { FunctionComponent, HTMLProps, ReactNode } from 'react';
 
 export interface DataListActionProps extends HTMLProps<HTMLDivElement> {
   'aria-labelledby': string;
   'aria-label': string;
   id: string;
+  actions: ReactNode[];
 }
 
 declare const DataListAction: FunctionComponent<DataListActionProps>;