+++ packages/patternfly-4/react-core/src/components/Wizard/WizardHeader.tsx
@@ -10,15 +10,17 @@ interface WizardHeaderProps {
   title: string;
   description: string;
   ariaLabel: string;
+  titleId: string;
+  descriptionId: string;
 }
 
-const WizardHeader: React.SFC<WizardHeaderProps> = ({ onClose, title, description, ariaLabel }) => (
+const WizardHeader: React.SFC<WizardHeaderProps> = ({ onClose, title, description, ariaLabel, titleId, descriptionId }) => (
   <div className={css(styles.wizardHeader)}>
     <Button variant="plain" className={css(styles.wizardClose)} aria-label={ariaLabel} onClick={onClose}>
       <TimesIcon aria-hidden="true" />
     </Button>
-    <Title size="3xl" className={css(styles.wizardTitle)}>{title}</Title>
-    {description && <p className={css(styles.wizardDescription)}>
+    <Title size="3xl" className={css(styles.wizardTitle)} aria-label={title} id={titleId}>{title}</Title>
+    {description && <p className={css(styles.wizardDescription)} id={descriptionId}>
       {description}
     </p>}
   </div>