+++ packages/patternfly-4/react-docs/src/components/docs/docs.styles.js
@@ -1,8 +1,0 @@
-import { StyleSheet } from '@patternfly/react-styles';
-import { global_spacer_xl as spacerXl } from '@patternfly/react-tokens';
-
-export default StyleSheet.create({
-  description: {
-    marginBottom: spacerXl.var
-  }
-});